refactor: domain package 생성
