+++ src/main/kotlin/blackjack/domain/Deck.kt
@@ -1,3 +1,3 @@
-package blackjack
+package blackjack.domain
 
 class Deck(val cards: Cards = Cards(Card.DECK))
