+++ src/test/kotlin/blackjack/CardTest.kt
@@ -1,5 +1,8 @@
 package blackjack
 
+import blackjack.domain.Card
+import blackjack.domain.Denomination
+import blackjack.domain.Suite
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
