+++ src/test/kotlin/blackjack/DeckTest.kt
@@ -1,5 +1,7 @@
 package blackjack
 
+import blackjack.domain.Card
+import blackjack.domain.Deck
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
