+++ src/test/kotlin/blackjack/DenominationTest.kt
@@ -1,5 +1,6 @@
 package blackjack
 
+import blackjack.domain.Denomination
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
