+++ src/test/kotlin/blackjack/SuiteTest.kt
@@ -1,5 +1,6 @@
 package blackjack
 
+import blackjack.domain.Suite
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
